Fix Space login routing and empty message state
