+++ web/apps/space/src/utils/spaceLoginNavigation.ts
@@ -1,5 +1,6 @@
 import type { NextRouter } from "next/router";
 import type { SetupProfile } from "screens/SetupProfileScreen";
+import { acceptPendingSpaceInvite } from "utils/spacePendingInvite";
 import { spaceRoutes } from "utils/spaceRoutes";
 
 type RefreshProfile = () => Promise<SetupProfile | null>;
@@ -9,7 +10,19 @@ export const routeAfterCompletedLogin = async (
     refreshProfile: RefreshProfile,
     mode: "push" | "replace" = "push",
 ) => {
-    await refreshProfile();
+    const profile = await refreshProfile();
+    if (profile) {
+        await acceptPendingSpaceInvite().catch((error: unknown) =>
+            console.error("Failed to accept pending invite", error),
+        );
+        if (mode == "replace") {
+            await router.replace(spaceRoutes.home);
+            return;
+        }
+        await router.push(spaceRoutes.home);
+        return;
+    }
+
     if (mode == "replace") {
         await router.replace(spaceRoutes.setupProfile("login"));
         return;
